Rename Timestamps to DateTimes
This class will contain functions for dealing with TIME type, too.
