+++ presto-main/src/main/java/io/prestosql/operator/scalar/FormatFunction.java
@@ -72,8 +72,8 @@ import static io.prestosql.spi.type.TinyintType.TINYINT;
 import static io.prestosql.spi.type.VarcharType.VARCHAR;
 import static io.prestosql.spi.type.Varchars.isVarcharType;
 import static io.prestosql.type.JsonType.JSON;
-import static io.prestosql.type.Timestamps.toLocalDateTime;
-import static io.prestosql.type.Timestamps.toZonedDateTime;
+import static io.prestosql.type.DateTimes.toLocalDateTime;
+import static io.prestosql.type.DateTimes.toZonedDateTime;
 import static io.prestosql.type.UnknownType.UNKNOWN;
 import static io.prestosql.util.Failures.internalError;
 import static io.prestosql.util.Reflection.methodHandle;
