+++ presto-main/src/main/java/io/prestosql/operator/scalar/timestamp/LocalTimestamp.java
@@ -20,17 +20,17 @@ import io.prestosql.spi.function.ScalarFunction;
 import io.prestosql.spi.function.SqlNullable;
 import io.prestosql.spi.function.SqlType;
 import io.prestosql.spi.type.LongTimestamp;
-import io.prestosql.type.Timestamps;
+import io.prestosql.type.DateTimes;
 
 import java.time.Instant;
 import java.time.LocalDateTime;
 import java.time.ZoneOffset;
 
 import static io.prestosql.spi.type.TimestampType.MAX_SHORT_PRECISION;
-import static io.prestosql.type.Timestamps.PICOSECONDS_PER_NANOSECOND;
-import static io.prestosql.type.Timestamps.epochSecondToMicrosWithRounding;
-import static io.prestosql.type.Timestamps.round;
-import static io.prestosql.type.Timestamps.scaleEpochMicrosToMillis;
+import static io.prestosql.type.DateTimes.PICOSECONDS_PER_NANOSECOND;
+import static io.prestosql.type.DateTimes.epochSecondToMicrosWithRounding;
+import static io.prestosql.type.DateTimes.round;
+import static io.prestosql.type.DateTimes.scaleEpochMicrosToMillis;
 
 @ScalarFunction(value = "$localtimestamp", hidden = true)
 public final class LocalTimestamp
@@ -75,6 +75,6 @@ public final class LocalTimestamp
                     .toInstant(ZoneOffset.UTC);
         }
 
-        return Timestamps.longTimestamp(precision, start);
+        return DateTimes.longTimestamp(precision, start);
     }
 }
