+++ presto-main/src/main/java/io/prestosql/operator/scalar/timestamp/SequenceIntervalDayToSecond.java
@@ -28,7 +28,7 @@ import static io.prestosql.operator.scalar.SequenceFunction.checkValidStep;
 import static io.prestosql.spi.type.TimestampType.MAX_SHORT_PRECISION;
 import static io.prestosql.spi.type.TimestampType.createTimestampType;
 import static io.prestosql.spi.type.TimestampTypes.writeLongTimestamp;
-import static io.prestosql.type.Timestamps.MICROSECONDS_PER_MILLISECOND;
+import static io.prestosql.type.DateTimes.MICROSECONDS_PER_MILLISECOND;
 import static java.lang.Math.multiplyExact;
 import static java.lang.Math.toIntExact;
