+++ presto-main/src/main/java/io/prestosql/operator/scalar/timestamp/TimestampToDateCast.java
@@ -26,7 +26,7 @@ import org.joda.time.chrono.ISOChronology;
 import java.util.concurrent.TimeUnit;
 
 import static io.prestosql.spi.function.OperatorType.CAST;
-import static io.prestosql.type.Timestamps.scaleEpochMicrosToMillis;
+import static io.prestosql.type.DateTimes.scaleEpochMicrosToMillis;
 import static io.prestosql.util.DateTimeZoneIndex.getChronology;
 
 @ScalarOperator(CAST)
