+++ presto-main/src/main/java/io/prestosql/operator/scalar/timestamp/TimestampToJsonCast.java
@@ -33,8 +33,8 @@ import java.time.format.DateTimeFormatter;
 import static io.prestosql.spi.StandardErrorCode.INVALID_CAST_ARGUMENT;
 import static io.prestosql.spi.function.OperatorType.CAST;
 import static io.prestosql.spi.type.StandardTypes.JSON;
-import static io.prestosql.type.Timestamps.formatTimestamp;
-import static io.prestosql.type.Timestamps.scaleEpochMillisToMicros;
+import static io.prestosql.type.DateTimes.formatTimestamp;
+import static io.prestosql.type.DateTimes.scaleEpochMillisToMicros;
 import static io.prestosql.util.JsonUtil.JSON_FACTORY;
 import static io.prestosql.util.JsonUtil.createJsonGenerator;
 import static java.lang.String.format;
