+++ presto-main/src/main/java/io/prestosql/operator/scalar/timestamptz/DateFormat.java
@@ -25,7 +25,7 @@ import io.prestosql.spi.type.StandardTypes;
 import static io.prestosql.operator.scalar.DateTimeFunctions.dateFormat;
 import static io.prestosql.spi.type.DateTimeEncoding.unpackMillisUtc;
 import static io.prestosql.spi.type.TimeZoneKey.getTimeZoneKey;
-import static io.prestosql.type.Timestamps.roundToEpochMillis;
+import static io.prestosql.type.DateTimes.roundToEpochMillis;
 import static io.prestosql.util.DateTimeZoneIndex.getChronology;
 import static io.prestosql.util.DateTimeZoneIndex.unpackChronology;
