+++ presto-main/src/main/java/io/prestosql/operator/scalar/timestamptz/TimeToTimestampWithTimeZoneCast.java
@@ -23,7 +23,7 @@ import io.prestosql.spi.type.StandardTypes;
 
 import static io.prestosql.spi.function.OperatorType.CAST;
 import static io.prestosql.spi.type.DateTimeEncoding.packDateTimeWithZone;
-import static io.prestosql.type.Timestamps.round;
+import static io.prestosql.type.DateTimes.round;
 import static io.prestosql.util.DateTimeZoneIndex.getChronology;
 
 @ScalarOperator(CAST)
