+++ presto-main/src/main/java/io/prestosql/operator/scalar/timestamptz/TimestampWithTimeZoneToTimeWithTimezoneCast.java
@@ -25,7 +25,7 @@ import static io.prestosql.spi.type.DateTimeEncoding.packDateTimeWithZone;
 import static io.prestosql.spi.type.DateTimeEncoding.unpackMillisUtc;
 import static io.prestosql.spi.type.DateTimeEncoding.unpackZoneKey;
 import static io.prestosql.spi.type.TimeZoneKey.getTimeZoneKey;
-import static io.prestosql.type.Timestamps.roundToEpochMillis;
+import static io.prestosql.type.DateTimes.roundToEpochMillis;
 import static io.prestosql.util.DateTimeZoneIndex.getChronology;
 
 @ScalarOperator(CAST)
