+++ presto-main/src/main/java/io/prestosql/sql/analyzer/ExpressionAnalyzer.java
@@ -169,11 +169,11 @@ import static io.prestosql.type.ArrayParametricType.ARRAY;
 import static io.prestosql.type.IntervalDayTimeType.INTERVAL_DAY_TIME;
 import static io.prestosql.type.IntervalYearMonthType.INTERVAL_YEAR_MONTH;
 import static io.prestosql.type.JsonType.JSON;
-import static io.prestosql.type.Timestamps.extractTimestampPrecision;
-import static io.prestosql.type.Timestamps.parseLegacyTimestamp;
-import static io.prestosql.type.Timestamps.parseTimestamp;
-import static io.prestosql.type.Timestamps.parseTimestampWithTimeZone;
-import static io.prestosql.type.Timestamps.timestampHasTimeZone;
+import static io.prestosql.type.DateTimes.extractTimestampPrecision;
+import static io.prestosql.type.DateTimes.parseLegacyTimestamp;
+import static io.prestosql.type.DateTimes.parseTimestamp;
+import static io.prestosql.type.DateTimes.parseTimestampWithTimeZone;
+import static io.prestosql.type.DateTimes.timestampHasTimeZone;
 import static io.prestosql.type.UnknownType.UNKNOWN;
 import static io.prestosql.util.DateTimeUtils.timeHasTimeZone;
 import static java.lang.Math.toIntExact;
