+++ presto-main/src/main/java/io/prestosql/sql/planner/LiteralInterpreter.java
@@ -53,9 +53,9 @@ import static io.prestosql.spi.type.VarcharType.VARCHAR;
 import static io.prestosql.sql.analyzer.SemanticExceptions.semanticException;
 import static io.prestosql.sql.analyzer.TypeSignatureProvider.fromTypes;
 import static io.prestosql.type.JsonType.JSON;
-import static io.prestosql.type.Timestamps.parseLegacyTimestamp;
-import static io.prestosql.type.Timestamps.parseTimestamp;
-import static io.prestosql.type.Timestamps.parseTimestampWithTimeZone;
+import static io.prestosql.type.DateTimes.parseLegacyTimestamp;
+import static io.prestosql.type.DateTimes.parseTimestamp;
+import static io.prestosql.type.DateTimes.parseTimestampWithTimeZone;
 import static io.prestosql.util.DateTimeUtils.parseDayTimeInterval;
 import static io.prestosql.util.DateTimeUtils.parseLegacyTime;
 import static io.prestosql.util.DateTimeUtils.parseTimeWithTimeZone;
