+++ presto-main/src/main/java/io/prestosql/sql/relational/SqlToRowExpressionTranslator.java
@@ -119,9 +119,9 @@ import static io.prestosql.sql.relational.SpecialForm.Form.ROW_CONSTRUCTOR;
 import static io.prestosql.sql.relational.SpecialForm.Form.SWITCH;
 import static io.prestosql.sql.relational.SpecialForm.Form.WHEN;
 import static io.prestosql.type.JsonType.JSON;
-import static io.prestosql.type.Timestamps.parseLegacyTimestamp;
-import static io.prestosql.type.Timestamps.parseTimestamp;
-import static io.prestosql.type.Timestamps.parseTimestampWithTimeZone;
+import static io.prestosql.type.DateTimes.parseLegacyTimestamp;
+import static io.prestosql.type.DateTimes.parseTimestamp;
+import static io.prestosql.type.DateTimes.parseTimestampWithTimeZone;
 import static io.prestosql.util.DateTimeUtils.parseDayTimeInterval;
 import static io.prestosql.util.DateTimeUtils.parseTimeWithTimeZone;
 import static io.prestosql.util.DateTimeUtils.parseTimeWithoutTimeZone;
